Move home page to root, remove unused views and users
The current live site shows the landing page with the contact form to subscribe to the news, including the header and footer which are in the layout view we use throughout the site.
This caused us to have to hide the navigation on this page only. Now for previewing the site we staged the V3 branch so we don't need to do that anymore so the following changes were made.

- Moved the home page to the root of the site.
- Removed the link to the homepage page from the navigation as it is included in the logo.
- Removed the footer and nav partials that were only displayed on the blog page.
-Also removed Statamic users that no longer exist.