Folded menu and layout rearrangement on mobiles
The layout that displayed on mobiles was rearranged and the widths of the columns in each section were modified.

The original side menu was hidden, and a new one was displayed exclusively on the small screens, as it must be folded out to use less space at the top of the page.