Fix error cause Prettier fail
The problems that cause Prettier to fail have been fixed, mostly because of errors in closing tags or because Prettier can't solve the most complex layouts.

_tile_steps, _features_icons
Replaced item counts that duplicated code with a single line of code that only modifies classes.

blog/index
The spacing that is added to articles at the top on the second page was added to the container to avoid adding a container only for the second page.

general_content/index
Comments that caused Prettier to crash were deleted.

_table, _breadcrumbs, _feature_section
Indentation error caused the error

author/index, _icon_features
Error in closing the container tag

Once the issues were resolved, Prettier was applied to these files.